monitoring_service/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
pika==1.3.2
python-dotenv==1.0.0
drf-spectacular==0.27.0
